requirements.txt
langchain
openai
chromadb
faiss-cpu
python-dotenv
PyPDF2
pypdf
tiktoken
streamlit
huggingface-hub
transformers
duckduckgo-search
pytest
langchain_community
langchain-google-genai
langchain_huggingface
fastapi 
uvicorn
requests
sentence-transformers
google-generativeai
python-dotenv
python-multipart
pydantic
firebase-admin
fpdf2